+++ packages/terra-table/tests/nightwatch/table-spec.js
@@ -65,7 +65,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
       .assert.elementPresent('tr[class*="is-selectable"]');
   },
 
-  'Display a selectable table and highlights the selected row upon clicking': (browser) => {
+  'Display a single selectable table and highlights the selected row upon clicking': (browser) => {
     browser.url(`${browser.launchUrl}/#/tests/table-tests/selectable-table`);
     browser.click('[class*="row"]:nth-child(1)');
     browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
@@ -83,7 +83,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
     browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.contains('is-selected');
   },
 
-  'Toggle aria-selected on seletable rows': (browser) => {
+  'Toggle aria-selected on single selectable rows': (browser) => {
     browser.url(`${browser.launchUrl}/#/tests/table-tests/selectable-table`);
     browser.click('[class*="row"]:nth-child(1)');
     browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('aria-selected').which.contains('true');
@@ -101,7 +101,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
     browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('aria-selected').which.contains('true');
   },
 
-  'Display a selectable table and highlights the selected row upon enter': (browser) => {
+  'Display a single selectable table and highlights the selected row upon enter': (browser) => {
     browser.url(`${browser.launchUrl}/#/tests/table-tests/selectable-table`);
 
     browser.sendKeys('[class*="row"]:nth-child(1)', browser.Keys.ENTER);
@@ -120,7 +120,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
     browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.contains('is-selected');
   },
 
-  'Display a selectable table and highlights the selected row upon space keydown': (browser) => {
+  'Display a single selectable table and highlights the selected row upon space keydown': (browser) => {
     browser.url(`${browser.launchUrl}/#/tests/table-tests/selectable-table`);
 
     browser.sendKeys('[class*="row"]:nth-child(1)', browser.Keys.SPACE);
@@ -139,7 +139,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
     browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.contains('is-selected');
   },
 
-  'Display a selectable table with a subheader': (browser) => {
+  'Display a single selectable table with a subheader': (browser) => {
     browser.url(`${browser.launchUrl}/#/tests/table-tests/table-selectable-subheaders`);
     browser.expect.element('tbody tr:nth-child(1)').to.have.attribute('data-terra-table-subheader-row');
     browser.assert.containsText('tbody tr:nth-child(1) td:nth-child(1)', 'Single');
@@ -155,7 +155,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
     browser.assert.containsText('tbody tr:nth-child(4) td:nth-child(1)', 'Married');
   },
 
-  'Triggers onChange for selectable table upon clicking a row': (browser) => {
+  'Triggers onChange for a single selectable table upon clicking a row': (browser) => {
     browser.url(`${browser.launchUrl}/#/tests/table-tests/selectable-table-onchange`);
 
     browser.click('[class*="row"]:nth-child(1)');
@@ -194,7 +194,7 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
     browser.assert.elementNotPresent('[class*="row"]:nth-child(2)');
   },
 
-  'Displays a selectable table with no rows': (browser) => {
+  'Displays a single selectable table with no rows': (browser) => {
     browser.url(`${browser.launchUrl}/#/tests/table-tests/selectable-table-no-rows`);
     browser.assert.elementNotPresent('[class*="row"]:nth-child(1)');
   },
@@ -207,5 +207,166 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
     browser.assert.elementPresent('.Cell-Custom');
     browser.assert.elementPresent('.SingleSelectableRows-Custom');
     browser.assert.elementPresent('.Subheader-Custom');
+  },
+
+  'Display a multiple selectable table with a subheader': (browser) => {
+    browser.url(`${browser.launchUrl}}/#/tests/table-tests/multiple-selectable-table-subheaders`);
+    browser.expect.element('tbody tr:nth-child(1)').to.have.attribute('data-terra-table-subheader-row');
+    browser.assert.containsText('tbody tr:nth-child(1) td:nth-child(1)', 'Single');
+    browser.expect.element('tbody tr:nth-child(4)').to.have.attribute('data-terra-table-subheader-row');
+    browser.assert.containsText('tbody tr:nth-child(4) td:nth-child(1)', 'Married');
+  },
+
+  'Displays a multiple selectable table with only one row': (browser) => {
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/multiple-selectable-table-single-row`);
+    browser.assert.elementPresent('[class*="row"]:nth-child(1)');
+    browser.assert.elementNotPresent('[class*="row"]:nth-child(2)');
+  },
+
+  'Displays a multiple selectable table with no rows': (browser) => {
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/multiple-selectable-table-no-rows`);
+    browser.assert.elementNotPresent('[class*="row"]:nth-child(1)');
+  },
+
+  'Triggers onChange for multiple selectable table upon clicking a row': (browser) => {
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/multiple-selectable-table-onchange`);
+
+    browser.click('[class*="row"]:nth-child(1)');
+    browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.assert.containsText('#selected-index', '[0]');
+
+    browser.sendKeys('[class*="row"]:nth-child(2)', browser.Keys.ENTER);
+    browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.assert.containsText('#selected-index', '[0,1]');
+
+    browser.sendKeys('[class*="row"]:nth-child(3)', browser.Keys.SPACE);
+    browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.contains('is-selected');
+    browser.assert.containsText('#selected-index', '[0,1,2]');
+  },
+
+  'Display a multiple selectable table and highlights the selected row upon enter': (browser) => {
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/multiple-selectable-table-multiple-rows`);
+
+    browser.sendKeys('[class*="row"]:nth-child(1)', browser.Keys.ENTER);
+    browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.not.contains('is-selected');
+
+    browser.sendKeys('[class*="row"]:nth-child(2)', browser.Keys.ENTER);
+    browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.not.contains('is-selected');
+
+    browser.sendKeys('[class*="row"]:nth-child(3)', browser.Keys.ENTER);
+    browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.contains('is-selected');
+  },
+
+  'Display a multiple selectable table and highlights the selected row upon space keydown': (browser) => {
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/multiple-selectable-table-multiple-rows`);
+
+    browser.sendKeys('[class*="row"]:nth-child(1)', browser.Keys.SPACE);
+    browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.not.contains('is-selected');
+
+    browser.sendKeys('[class*="row"]:nth-child(2)', browser.Keys.SPACE);
+    browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.not.contains('is-selected');
+
+    browser.sendKeys('[class*="row"]:nth-child(3)', browser.Keys.SPACE);
+    browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.contains('is-selected');
+  },
+
+  'Display a multiple selectable table and removes highlight from already selected row upon click': (browser) => {
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/multiple-selectable-table-multiple-rows`);
+
+    browser.click('[class*="row"]:nth-child(1)');
+    browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.not.contains('is-selected');
+
+    browser.click('[class*="row"]:nth-child(1)');
+    browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.not.contains('is-selected');
+  },
+
+  'Display a multiple selectable table and removes highlight from already selected row upon enter': (browser) => {
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/multiple-selectable-table-multiple-rows`);
+
+    browser.sendKeys('[class*="row"]:nth-child(1)', browser.Keys.ENTER);
+    browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.not.contains('is-selected');
+
+    browser.sendKeys('[class*="row"]:nth-child(1)', browser.Keys.ENTER);
+    browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.not.contains('is-selected');
+  },
+
+  'Display a multiple selectable table and removes highlight from already selected row upon space keydown': (browser) => {
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/multiple-selectable-table-multiple-rows`);
+
+    browser.sendKeys('[class*="row"]:nth-child(1)', browser.Keys.SPACE);
+    browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.not.contains('is-selected');
+
+    browser.sendKeys('[class*="row"]:nth-child(1)', browser.Keys.SPACE);
+    browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.not.contains('is-selected');
+  },
+
+  'Displays a multiple selectable table with a preselected row': (browser) => {
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/multiple-selectable-table-preselected-row`);
+
+    browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.click('[class*="row"]:nth-child(1)');
+    browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
+
+    browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.click('[class*="row"]:nth-child(2)');
+    browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.not.contains('is-selected');
+
+    browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.click('[class*="row"]:nth-child(3)');
+    browser.expect.element('[class*="row"]:nth-child(3)').to.have.attribute('class').which.contains('is-selected');
+
+    browser.expect.element('[class*="row"]:nth-child(4)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.click('[class*="row"]:nth-child(4)');
+    browser.expect.element('[class*="row"]:nth-child(4)').to.have.attribute('class').which.contains('is-selected');
+
+    browser.expect.element('[class*="row"]:nth-child(5)').to.have.attribute('class').which.contains('is-selected');
+    browser.click('[class*="row"]:nth-child(5)');
+    browser.expect.element('[class*="row"]:nth-child(5)').to.have.attribute('class').which.not.contains('is-selected');
+
+    browser.expect.element('[class*="row"]:nth-child(6)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.click('[class*="row"]:nth-child(6)');
+    browser.expect.element('[class*="row"]:nth-child(6)').to.have.attribute('class').which.not.contains('is-selected');
+
+    browser.expect.element('[class*="row"]:nth-child(7)').to.have.attribute('class').which.not.contains('is-selected');
+    browser.click('[class*="row"]:nth-child(7)');
+    browser.expect.element('[class*="row"]:nth-child(7)').to.have.attribute('class').which.contains('is-selected');
+
+    browser.expect.element('[class*="row"]:nth-child(8)').to.have.attribute('class').which.contains('is-selected');
+    browser.click('[class*="row"]:nth-child(8)');
+    browser.expect.element('[class*="row"]:nth-child(8)').to.have.attribute('class').which.contains('is-selected');
+
+    browser.expect.element('[class*="row"]:nth-child(9)').to.have.attribute('class').which.contains('is-selected');
+    browser.click('[class*="row"]:nth-child(9)');
+    browser.expect.element('[class*="row"]:nth-child(9)').to.have.attribute('class').which.not.contains('is-selected');
   },
 });